requirements.txt
celery
Django
djangorestframework
requests
django-cors-headers
